+++ lib/src/exports/extensions/flutter-widgets/build_ctx.dart
@@ -321,9 +321,26 @@ extension Buildctx on BuildContext {
     OverlayEntry? above,
     OverlayEntry? below,
   }) {
-    OverlayEntry overlayEntry =
-        OverlayEntry(builder: (context) => overlayWidget);
-
+    final overlayEntry = OverlayEntry(builder: (context) => overlayWidget);
     Overlay.of(this).insert(overlayEntry, above: above, below: below);
   }
+
+  /// The function `dialogBx` displays a dialog box with the provided widget.
+  ///
+  /// Args:
+  ///   box (Widget): The `box` parameter is a required `Widget` that represents the content of the
+  /// dialog box. It can be any widget that you want to display inside the dialog box.
+  Future<void> dialogBx({required Widget box}) => showDialog(
+        context: this,
+        builder: (context) => box,
+      );
+
+  /// The function `botmSht` is a Dart function that displays a modal bottom sheet with the provided
+  /// widget.
+  ///
+  /// Args:
+  ///   btmSht (Widget): The `btmSht` parameter is a required `Widget` that represents the content of
+  /// the bottom sheet. It will be displayed inside the `showModalBottomSheet` function.
+  Future<void> botmSht({required Widget btmSht}) =>
+      showModalBottomSheet(context: this, builder: (context) => btmSht);
 }